feat: geocoding
Geocoding active for both Events and Reports. Events, when registered, need the address to be properly put. This address is converted to lat, lng using the mapbox geocoding function. These are then stored on the database as well. Reports, when sent, ask the user to provide their location. Without it, no reporting can be done. This is converted to lat lng using browser geolocation and similarly sent to the db